requirement.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.8"
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-extensions==3.2.3
django-jazzmin==2.6.0
et-xmlfile==1.1.0
idna==3.4
kaleido==0.2.1
mysqlclient==2.2.0
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.17.0
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
six==1.16.0
smartapi-python==1.3.5
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
websocket-client==1.6.3
